backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
django-filter==24.3
psycopg2-binary==2.9.9
WeasyPrint==62.3
whitenoise==6.7.0
Pillow==12.0.0
django-cors-headers==4.5.0
openpyxl==3.1.5
pandas==2.2.3
pyotp==2.9.0
qrcode==7.4.2
python-telegram-bot==20.8
